[Behat][Admin] Cover order history tests
