+++ features/admin/order/managing_orders/order_history/tracking_changes_on_order_addresses.feature
@@ -13,7 +13,7 @@ Feature: Tracking changes on order addresses
         And the customer bought a single "Italian suit"
         When I am logged in as an administrator
 
-    @api @todo @ui
+    @api @ui
     Scenario: Browsing order's addresses history after changing it by customer
         Given the customer "Barney Stinson" addressed it to "East 84st Street and 3rd Avenue", "10118" "New York" in the "United States" with identical billing address
         And the customer changed shipping address' street to "211 Madison Avenue"
@@ -22,7 +22,7 @@ Feature: Tracking changes on order addresses
         Then there should be 2 shipping address changes in the registry
         Then there should be 1 billing address changes in the registry
 
-    @api @todo @ui
+    @api @ui
     Scenario: Browsing order's addresses history after changing it by administrator
         Given the customer "Barney Stinson" addressed it to "East 84st Street and 3rd Avenue", "10118" "New York" in the "United States" with identical billing address
         And the customer chose "Free" shipping method with "Cash on Delivery" payment
