+++ src/Sylius/Behat/Resources/config/suites/ui/order/order_history.yml
@@ -25,10 +25,10 @@ default:
                 - sylius.behat.context.transform.product
                 - sylius.behat.context.transform.shipping_method
                 - sylius.behat.context.transform.zone
-
                 - sylius.behat.context.transform.shared_storage
 
                 - sylius.behat.context.ui.admin.managing_orders
+                - sylius.behat.context.ui.admin.order_history
 
             filters:
                 tags: "@order_history&&@ui"
